fix(docker-tests): eliminate tendermint and integration feature warnings
- Move resolve_compose_container_id from docker_ops.rs to env.rs so
  tendermint can use it without needing the docker_ops module
- Remove docker-tests-tendermint from docker_ops module gate
- Gate serde_json macro_use on features that need it (tendermint tests
  have explicit `use serde_json::json` imports)
- Remove docker-tests-integration from block_on import (only used by setup_sia)
- Remove docker-tests-integration from UtxoAssetDockerOps import (only
  used by setup_utxo which isn't called for integration)
- Add docker-tests-integration to SET_BURN_PUBKEY_TO_ALICE import in swap.rs
